Move aliases into alias.c, move non-aliases into nvram.c
